Move DatafusionContext to session.rs
