+++ pg_analytics/src/datafusion/mod.rs
@@ -4,4 +4,5 @@ pub mod context;
 pub mod datatype;
 pub mod directory;
 pub mod schema;
+pub mod session;
 pub mod table;
