+++ pg_analytics/src/hooks/drop.rs
@@ -1,7 +1,7 @@
 use deltalake::datafusion::catalog::schema::SchemaProvider;
 use pgrx::*;
 
-use crate::datafusion::context::DatafusionContext;
+use crate::datafusion::session::DatafusionContext;
 use crate::errors::ParadeError;
 use crate::hooks::handler::IsColumn;
